feat, refactor: add tests for NominatimService, adapt dependent tests
